extract deliverEntityActivitiesFromPatch to its own file
