+++ server/controllers/activitypub/lib/radio_hooks.js
@@ -4,10 +4,9 @@ const radio = require('lib/radio')
 const user_ = require('controllers/user/lib/user')
 const { postActivityToActorFollowersInboxes } = require('./post_activity')
 const { byActorName, createActivity } = require('./activities')
-const formatEntityPatchesActivities = require('./format_entity_patches_activities')
 const formatUserItemsActivities = require('./format_user_items_activities')
 const formatShelfItemsActivities = require('./format_shelf_items_activities')
-
+const deliverEntityActivitiesFromPatch = require('./deliver_entity_activities_from_patch')
 const debouncedActivities = {}
 
 module.exports = () => {
@@ -20,7 +19,7 @@ module.exports = () => {
   radio.on('shelves:update', shelvesIds => {
     return Promise.all(shelvesIds.map(debounceActivities))
   })
-  radio.on('patch:created', _deliverEntityActivitiesFromPatch)
+  radio.on('patch:created', deliverEntityActivitiesFromPatch)
 }
 
 const createDebouncedActivity = ({ userId, shelfId }) => async () => {
@@ -62,45 +61,4 @@ const debounceActivities = async shelfId => {
     debouncedActivities[shelfId] = _.debounce(createDebouncedActivity({ shelfId }), activitiesDebounceTime)
   }
   return debouncedActivities[shelfId]()
-}
-
-const _deliverEntityActivitiesFromPatch = patch => {
-  return deliverEntityActivitiesFromPatch(patch)
-  .catch(_.Error('create_activities_on_entities_updates err'))
-}
-
-const deliverEntityActivitiesFromPatch = async patch => {
-  const rows = byClaimValueAndDate(patch)
-  if (rows.length === 0) return
-  const activities = await formatEntityPatchesActivities(rows)
-  if (activities.length === 0) return
-  await Promise.all(activities.map(activity => {
-    const actorName = new URL(activity.actor).searchParams.get('name')
-    return postActivityToActorFollowersInboxes({ activity, actorName })
-  }))
-}
-
-// Mimick server/db/couchdb/design_docs/patches.json byClaimValueAndDate
-const byClaimValueAndDate = doc => {
-  const { _id: id } = doc
-  const rows = []
-  for (const operation of doc.patch) {
-    if (operation.op === 'add') {
-      const [ , section, property, arrayIndex ] = operation.path.split('/')
-      if (section === 'claims') {
-        if (arrayIndex != null) {
-          if (operation.value.startsWith('wd:') || operation.value.startsWith('inv:')) {
-            rows.push({ id, key: [ operation.value, doc.timestamp ], value: property })
-          }
-        } else {
-          for (const subvalue of operation.value) {
-            if (subvalue.startsWith('wd:') || subvalue.startsWith('inv:')) {
-              rows.push({ id, key: [ subvalue, doc.timestamp ], value: property })
-            }
-          }
-        }
-      }
-    }
-  }
-  return rows
 }
